Tests/LibWeb/Text/expected/Streams/ReadableStreamBYOBReader-read-min.txt

This -⌄
 will not be processed as one chunk
This -> will be processed as one chunk
Stream closed
